Update LDAP.pm
CHG: if you use a password with a colon, you cannot log into the ldap account, because split() will destroy your password. the change will limit the pieces to two, so we have part one, the username and part two the password with colon. I'm not sure, whether or not it is possible to have an username with colons...